requirements.txt
requests
requests-cache
gitpython
keyring
